# Includes HABTM returns correct size now