@scambra 2015-02-05
Is there any issue preventing this to be merged? I'm stuck on rails 4.0 because of this bug and it would be great to get this fixed in next 4.1 and 4.2 releases