@hjdivad 2015-05-13
@sgrif @eileencodes did you have a chance to review this?  Is there something that can be done to move this PR forward?  It's causing quite a lot of problems for us and we'd love to get the issue fixed in rails 4.1 or 4.2.

@scambra I tested this in our app and it fixed the problem for us.  Unfortunately it seems to have a performance regression, at least as compared to a manual preloading workaround (which is what we currently do).  In our app the performance regression is almost an order of magnitude.  I have reproduced a performance regression in a simpler app, although here the difference is only 27%.  I suspect the reason has to do with the size of the rows coming back from the db, but this is pure speculation on my part.


The sample app is at hjdivad/rails-18383-patchtest.  the `master` branch is using `4-1-stable` and the branch `patched` is your commit applied.  You can see the [diff here](https://github.com/hjdivad/rails/compare/4-1-stable...hjdivad:patch-4.1).

On master in that branch, on my machine (a mid 2012 macbook pro) I get the following results:

```
# on master
$ rake test:benchmark SEED=62777
Run options: --seed 62777

# Running:

M2MTest#test_homepage (523 ms warmup)
           wall_time: 140 ms
              memory: unsupported
             objects: unsupported
             gc_runs: 0

             gc_time: 0 ms
.

Finished in 2.895263s, 0.3454 runs/s, 4.4901 assertions/s.

1 runs, 13 assertions, 0 failures, 0 errors, 0 skips


# on patched
$ rake test:benchmark SEED=62777
Run options: --seed 62777

# Running:

M2MTest#test_homepage (561 ms warmup)
           wall_time: 195 ms
              memory: unsupported
             objects: unsupported
             gc_runs: 0
             gc_time: 0 ms
.

Finished in 3.650158s, 0.2740 runs/s, 3.5615 assertions/s.

1 runs, 13 assertions, 0 failures, 0 errors, 0 skips
```

The manual preloading that I'd love this PR to replace can be found [here](https://github.com/hjdivad/rails-18383-patchtest/blob/master/app/models/war.rb#L14-L47).