@scambra 2015-05-20
@hjdivad Can you check if preload sql is using a slower sql than your workaround?