@eileencodes 2015-05-20
We should use https://github.com/evanphx/benchmark-ips to measure whether there is a significant slow down or not.