@scambra 2015-05-21
@hjdivad I think you should remove "load_participants! wars" line on patched branch, because participants are eager loaded.