@hjdivad 2015-05-21
@scambra you are 100% correct.  I have the line commented out locally when I ran the benchmarks.  Sorry I'd apparently not pushed the latest from this branch; I have done so now.

I'll try to rerun and catch the SQL to see if we can learn anything interesting about what it generates.  One thing I noticed when looking at the graphs generated by tmm1/perftools.rb was a lot of `Hash#hash` and garbage collection.

In the patch we key on entire rows returned by the join when only a tiny subset is necessary.  This happens [here](https://github.com/scambra/rails/blob/33a5416461bc5068d85d33259310f56625569e86/activerecord/lib/active_record/associations/join_dependency.rb#L236).

I changed that to `slice` the primary and foreign keys and that cut out some of the cost in our app.  It had much less effect in the patchtest app, which is unsurprising because the rows returned in the join there are much smaller.

I don't understand the data structures here enough to know how to properly extract the primary and foreign keys: I just hardcoded them for my app when testing.