@rafaelfranca 2015-08-14
@hjdivad @scambra Hey guys, can we confirm that this patch caused a performance regression? I'm planing to merge but I don't want to introduce an regression.