@rafaelfranca 2015-08-14
BTW, I'm releasing a new release candidate tomorrow and if we don't fix this issue tomorrow I'll push it one more release.