@scambra 2015-08-18
I think it's a performance regression with @hjdivad custom code (performance regression, at least as compared to a manual preloading workaround), not a rails regression because with rails it doesn't work.

So rails is slower than this [workaround](https://github.com/hjdivad/rails-18383-patchtest/blob/master/app/models/war.rb#L14-L47)