@hjdivad 2015-09-02
Hey, sorry I haven't had time to look at this more closely.  FWIW I'm :+1: on merging.  As @scambra says, in current rails this feature just doesn't work: worse, it appears to work and will work for the first record, so it's quite a landmine.

If I had more time, the comparison to make would be this PR against the last version of rails in which the feature was not broken.  I'm not actually sure that this PR is slower than the feature was in earlier versions of rails: I was wrong to use the word "regression" above: what I meant was not a regression of rails per se (as I haven't compared it to an earlier version of rails) but a regression in our app when swapping out our manual preloading for this patched version of rails.

As @scambra and I were discussing above, this PR is nontrivially slower than manually preloading.  From what I could tell when I looked into this earlier, at least some portion of the cost comes from hashing many large result sets, when the only thing that needs to be hashed are the primary and foreign keys.