@scambra 2015-09-29
@rafaelfranca are you going to merge? @hjdivad and me have confirmed there is no regression, I would like to upgrade to next version