@sgrif 2015-10-20
Can you please post the benchmark you used to confirm that there is no regression so we can verify?