@scambra 2015-10-21
I didn't benchmark any regression, I don't understand what regression you talk about, right now it doesn't work, so I don't think a regression it's possible. It doesn't affect queries working, only affects to preloading habtm, which doesn't work if they are preloaded without left outer join.

@hjdivad talked about regression, but he already said he was wrong using word regression, because he was comparing with his manual preloading, not comparing with rails