@hjdivad 2015-10-21
@sgrif yeah sorry for the confusion.  The comparisons I posted earlier compare [manual preloading](https://github.com/hjdivad/rails-18383-patchtest/blob/master/app/models/war.rb#L14-L47) with this patched version of rails.  This suggestions there is nontrivial room for improvement, but it's not a reason to believe that there's actually a regression in rails.

Our app has always had this problem because we've never been on an old enough version of rails where this kind of eager loading actually worked.