@rafaelfranca 2015-10-28
Backported in ec77ed7 and 5508474